+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 
+### Added
+- Added feature to install custom firmware from file ([pybricks-code#1020]).
+
 ### Fixed
 - Fixed run button enabled when no file open ([support#691]).
 - Fixed flash firmware dialog not showing when settings not open ([support#694]).
@@ -11,6 +14,7 @@
 - Fixed imports with invalid file name silently ignored ([support#717]).
 
 [pybricks-code#1011]: https://github.com/pybricks/pybricks-code/issues/1011
+[pybricks-code#1020]: https://github.com/pybricks/pybricks-code/issues/1020
 [support#691]: https://github.com/pybricks/support/issues/691
 [support#694]: https://github.com/pybricks/support/issues/694
 [support#717]: https://github.com/pybricks/support/issues/717